Remove unused parameter from KoInlineObject::updatePosition
The QTextObject parameter became obsolete when our KoInlineObjects
stopped being QTextObjects. We now figure out which inline object we
encounter through the char format.

However, the QTextObject parameter was still passed to all KoInlineObject's
updatePosition methods and that was confusing.

Note: maybe we should remove it from resize and paint as well.