Index marker: insert using the shape selection, not the tool proxy
Using KoToolProxy->selection and casting that into a KoTextEditor is is
confusing: it is meant to allow access to the selection, not to allow
access to the complete contents of a shape.

Now we get the selected shape from the active canvas and if it has the
right userdata, get the right KoTextEditor from it, and use that to
insert the variable.

Also, KoTextEditor now return the variable instead of a bool, because
the InsertInlineObjectActionBase::createInlineObject wants to return
that.