TextPasteCommand: don't switch to the text tool just to get a KoTextEditor
The RDF code in TextPasteCommand::redo() called ensureTextTool to switch
to the text tool and then get the KoTextEditor through the KoToolSelection
from the text tool, while all the while we already have a KoTextEditor
on the current document -- which is, in fact, the same object we got
through ensureTextTool.

So re-use that editor.