Be sure we always set the QTextBlock and always regenerate the toc 2 times at the beginning. This reintroduces the generate() during layout-run logic which we indeed need to minimize the numbers of relayouts needed. That reopens the original problem we had with headers that are above the toc may result in us dealing with the wrong QTextBlock cause the position() changed then. Will look how to solve that in a better way.
Also this still does not fix the problem that those
The_Miscellaneous_Icons_And_Scroll_Bars.odt produces wrong anchor-position results
cause of the toc's relayout. But for that we probably need to repair the way how
anchor repositioning is done atm (to be more exact it's not done at all).