Relax the update condition
We should be able to update even when the image is locked: the updates
are queued in such cases.

The only drawback of such relaxation is that some old code calls setDirty
insane number of times, e.g. KisColorSpaceConvertVisitor. This should
be fixed by porting this visitor to 'processing' framework.