Change key/value types around for KoDocumentRdf's KoTextInlineRdf map
We mapped from object to xmlid, but we retrieved the object xmlid.

Also, go back to using weak pointers, as already noted in the dox to make
sure we know when the KoTextInlineRdf objects get deleted. If we encounter
a deleted object, we remove it from the list.

This replaces the connection to KoTextInlineRdf's destroyed signal
which didn't work.