Textshape: Make the style manager work again.
After this patch, when you double click on one of the styles in the
listview to the left, you will get a tabbed dialog to the right with
details about the style.

This patch is the first from Mojtaba Shahi Senobari (moji).  Great job!

Removing of the canDelete parameter was discussed with Pierre Stirnweiss.