Use a stack of stacks for the inline object start/end matching
Previously, if you had a document that contained a table with a bookmark
start before and bookmark end after the table, all table cells would
be written out with a bookmark-end because we call writeblocks recursively,
even on nested documents.

This patch creates a stack of stacks which gets pushed at the start
of writeBlocks and popped at the end of writeblock. Hopefully, there's
not more KoTextWriter::Private state that gets confused because we
call writeBlocks recursively.