Make KoTextEditor::cursor private and KoTextEditor::document const
After discussion on irc, we agreed that KoTextEditor should be the single
facade that hides all editing of QTextDocuments. This means that the
cursor() method should be private and no longer used (except for now in
the commands that KoTextEditor creates) and that the QTextDocument
pointer must be const.

This also removes the QTextCursor based findXXX() methods in KoDocumenRdf.
Because the other findXXX(KoTextEditor) methods in KODocumentRdf depended on getting
a QTextCursor from KoTextEditor, those had to be rewritten using
QTextDocument::find() -- which should be faster than iterating over
characters anyway, and there isn't a limit of 500 characters from the
search start point anymore either.