Houston we have a problem! Qt can only handle one set of tabs per paragraph and that just doesn't cut it for the combination of tabs and first line indent (aka textIndent) as well as the combination of tabs and run around.
The reason is that Qt tabs start with the first letter of the line so when we move a line around
because of indentation or run around we move the tabs as well. And the convention is that tabs
should stay fixed in respect to textarea (possibly relative to margin).

Anyway all we can do is make tabs work on the first line even with textIndent, but on following
lines tabs will be placed wrongly and tabs also don't work if there is run around.

The solution is to patch QTextLine to take a tabsOffset.