Connected scheduler's queues to the global progress bar
From now on the use of KisProgressUpdater created by KisProgressInterface
should be deprecated, because the progress of all the actions is displayed
by the scheduler automatically. You should just split your task into
several jobs. To show the progress of a single task you can still use
the progress bars in the layer box.