Add method to get draw styles of a defined type
This patch makes it possible to get all gradients, markers, ...  from the styles reader.
It is needed when we don't want to load the same stuff over and over again.
I will use it in my marker branch to get all markers and load them in one go and just
reuse them later.

There is no change in functionality for the extiting applications.

REVIEW: 102562