Add the ability to use previews for frames with unsupported formats.
It will show the embedded objects in order of preference. If the
first object is not recognised, the new Unavail shape will be used,
which will then try the other options in turn, and for each attempt
to load a Picture shape, and upon success it will load the data and
disable the Picture shape's tools (to avoid confusion etc). Finally,
if no object is recognised, a question mark image is shown, to
indicate that the object is unknown.

Work done by Inge Wallin and Dan Leinir Turthra Jensen.

REVIEW:102512