Do not pass a KoResourceManager to KoTextEditor
The resourcemanager we actually need (for rdf) is the document resource
manager. This we need to get from the shapeController, which we also
pass. So, remove the KoResourceManager parameter since we might be confused
and pass the canvas resourcemanager.

Note: Thorsten and Casper we completely right when they found having
a KoResourceManager both in the document as well as in the canvas really
confusing. Yes, a key-value map with some predefined values is a very
basic thing and it feels wrong to have more than one implementation of it,
but it's really evil to have two variables with the same type and name
actually do completely different things and never mark the actual use...

I consider this a design bug created in January 2010, which we really
should fix.

See http://lists.kde.org/?l=koffice-devel&m=126483273211844&w=2