be more efficient when saving generic shapes to svg
when saving shapes without an own saveSvg method to svg
we now first try to paint the shape using an svg generator.
If that works we embed the generated svg, otherwise we create
and image of the shape and save that.
before we did both which did increase the resuling file size
unneccessarily.