Separate KoResourceManager for canvas and document
KoResourceManager was used in two places: the canvas and document
classes. This gave rise to endless confusion about which resource
manager should be used where and contains what. The KoResourceManager
api also contained methods that were only relevant when used from
a document and others that were only relevant when used from a canvas.

This patch splits the resourcemanager into two public and one private
class. The public classes are KoDocumentResourceManager and
KoCanvasResourceManager, the private class is KoResourceManager. The
public classes defer to the private class for code-sharing purposes.

The old KoCanvasResource and KoDocumentResource enums have been taken
into the classes themselves.