Fixed the last race in COW mechanism
After the COW we shouldn't release old tile data instantly, because
some of the other lockers of the tile might still access it. That's
why we now store all the disposed tiles in a special stack and free
them right in the moment when there are no lockers left.

Actually, both Valgrind and Intel Inspector still report a couple of races
there, but they are wrong. They think that the constructor of the cloned
tile data conflicts with readers. But that is not so. Cloning
(construction of the data) and any further reading accesses are separated
with the line kis_tile.cc:188:

m_tileData = tileData;

It means that the cloned tile is inaccessible for readers/writers during
construction. So there is no race condition possible.

BUG:273908