Make the location of the swapfile configurable
I do not intend to expose this from the settings gui, it's too advanced
for that. Advanced users can use the "swaplocation" entry in kritarc
to move the swapfiles out of the temp dir.

For the swapfile, we now use KTemporaryFile instead of QTemporaryFile
since we're still a KDE application.

BUG:131930