Fixed a race condition bug in KoResourceLoaderThread
There was a bug: the users of the server provider might request
a server before the thread has loaded it. That caused deleting
the thread object with all its consequences. No that is gone.

We call to KoResourceLoaderThread::barrier() before accessing
the server itself. This call is quite cheap (mutex+read) when
the thread has already finished running, and when the loading
is not completed yet, the caller is blocked until it is done.