Fix the inheritance of paragraph styles  before we loaded every inherited value into the child style too
Make ParagraphStyle a subclass of KoCharacterStyle so we handle the
char properties as one style and not auxillary style.

We also no longer add the character properties to the stylemanager as
a character style of it's own. And thus we no longer need the hack
that prevented them from being seen by the user in StylesModel.

Let characterstyles have inheritance too
This proves to be a big improvement getting font's right in a lot of files