Final border refactoring
This finally removes KoTableBorderStyle.

cstester does not report issues anymore : the issue was not directly related
to that change, we had (again) overlapping properties between KoTableCellStyle
and the KoParagraphStyle, thus borders could be dropped since they are applied
before the paragraph style...

This also contains a heavy rewrite of the getters/setters in KoBorder
in order to simplify its code.

Got a "ship it" by boemann during the sprint.

REVIEW: 102998