use clicked, not activated to select a brush engine
If we use the activated() signal to select a paintop, the user will be
able to select a paintop in the list with a single click on KDE -- but
on double-click platforms, like Windows, they will have to double-click.

That's not expected behaviour for items in a list, so use the simple
clicked() signal.