Do not reset custom shortcuts for tool actions on tool switch
If you had set a custom shortcut for an action owned by a tool, it would
get reset to the default value on switching to another tool and back.

This is because we explicitly set the shortcut to the default one; so
let's not do that.

Note: this patch also chances several instances of QAction to KAction.
It's dangerous to use the base QAction class when dealing with a KAction
because the setShortCut method is _not_ virtual.

See: http://api.kde.org/4.x-api/kdelibs-apidocs/kdeui/html/classKAction.html#_details