Added checks for 'zero'ness of the scale of the brush
When using tablet, the last events coming from it are almost zero,
so the resulting scale is zero as well. Default mask generator will get
crazy if it gets zero scale due to division by zero. So we need to check
in most of the tools whether the scale is not zero.

BUG:289145