Save fo:font-family with attributes that need it
This fixes bug 289980 -  Some font info lost on roundtrip cycle

The problem was caused due to the following problem

The paragraph contains fo:font-family and fo:font-family-generic with roman
Text in the paragraph contains the same fo:font-family but a different fo:font-family-generic.
The removal of the duplicate entries removed the fo:font-family and so only the
fo:font-family-generic was saved. However according to the odf spec these value is only loaded
if a fo:font-family is there in the same text-properties. As it was not there on loading we
ran in the problem described in the above bug report.
Now we make sure that in removeDuplicates the fo:font-family is kept when any of the properties
that needs it for loading is there after the duplicates have been removed.

BUG: 289980