Only show the GHNS upload button if the attica version is high enough
Attica is buggy prior to 3.0 (a.k.a. 2.9) because of bug 281850. This
caused Krita bug 285441, where upload would fail sometimes because of
missing content-id. There's nothing we can do to fix that from within
Krita, so only show the upload button if the version of attica is new
enough. The download button is always enabled.

BUG: 285441
CCBUG: 281850