Don't write invalide font-pitch value
This fixes writing of invalid odf as the font-pitch was saved into the paragraph-properties.
This reverts parts of commit b2d2d260141b1516cb897a82c7db4e144845a7d7 that handle font-pitch loading.
The font-pitch is already loaded into QTextFormat::FontFixedPitch. Please not that according to the odf
spec this property is only loaded when in the same text-properties also a fo:font-family is given.

CCMAIL: pinaraf@pinaraf.info