Fix whitespace handling in xml loading
This changes the xml loading to no longer strip whitespace. This fixes one of the
problems attached to the bug report where there is a text node with only a space
e.g. <text:line-break/> <text:s/>2 space.
This change uncovered some problems where white space was not handled properly:
anchors and line-breaks in filters, frame element loading in libs.
It also fixes saving of line-break to use the correct tag instead of the QChar::LineSperator as char.
Also whitespace handling after tabs, bookmark-ref and when variables are used have been fixed.

REVIEW: 103568

BUG: 290091