Delete KoTextAnchor from object manager on delete
Add a way to add object to the KoInlineTextObjectManager without adding them
to the document. This is needed so it is possible to use it in commands e.g.
the DeleteAnchorsCommand that removes the anchor once the replacement char
has been removed from the document and adds it again on undo. That should fix
the problem that the KoInlineTextObjectManager still knows about objects no
longer being around.