Added classes for dummy representation of the nodes in UI context
The highlevel interface of this representation is KisNodeShapesGraph.
It follows the hierarchy of nodes and arranges internal dummy nodes
accordingly.

Currently it is not fully connected to the KisShapeController. There
not connection to node move signals. But in this stage it fully resembles
old behavior and there should be no regressions.