Implement tripleclicking To do so I had to reroute doubleclicking into toolproxy mousepress
However for applications there are no change, except tools can now
reimplement mouseTripleClickEvent to be notified of a triple click

4 or more clicks are interpreted as (repeated) tripleclick. That
is until the doubleclicktimeout period has elapsed since the first
click.

The behaviour is smooth and as expected. I've also change texttool
take advantage of it and actiually select the paragraph on third
click and not as before where doubliclcking alternated between
selecting word and line

This patch also fixes a bug where doublclickig a word selects the
space after (it's a matter of taste and windows always selects the
space. but it seems most people like to only select the word.
BUG: 291124