Calculate the selectionBoundingBox correctly
In case the end of the selection was in a different root area the selection area
was not updated correctly. Update the end of the selection bounding box if the
full paragraph is selected.

This fixes bug 291413 -  Delecting a selection that goes over >1 page leaves part of it still looking selected

BUG: 291413