Don't loose the InlineInstanceId on apply style
We need to keep the InlineInstanceId when we apply a style. Otherwise we can no longer
find the inline object to the replacement character. This e.g. results in problems with
bookmark as the bookmark was no longer found if you changed the style of the text containing
the bookmark and after that changed the position of the bookmark.