Make PhongBumpmap iterators const
BUG:289190
o The iterators were formerly non-const because when the filter was made they were broken. Making them const now corrected the artifacts on the borders of 512x512 tiles when applying the filter.
o changedRect now simply returns 'rect'. This hasn't changed the observed behavior of the filter but is more correct
o Change the lines about inputArea and outputArea for clarity, these variables still retain the same values, but they are now clearer.