Fixed an order of destruction of KisCanvas2 and KisCanvasController
Before:
KisCanvas2 was deleted in the destructor of KisView2
KisCanvasController was deleted by QObject destructor of KisView2

It caused a problem: KisCanvasControlled called docker::unsetCanvas()
after the canvas has actually been deleted. That caused crashes.

After:
Now both of them are deleted in the destructor of KisView2, so that
we explicitly define the order of their destruction.