Improve anchoringParagraphRect top
For anchoring the top of the rect is supposed to be the point where
potential borders are drawn too.
Previously we used the point where background was filled. Without
borders that was not the correct place, so now always use the
"divider" position.

In user terms it means top is above the letters where the paragraph
spacing is counted from.