Fix filling shapes bug
Now tools add the stroke/filling styles to the resources snapshot
and it setups the painter in right way.

I removed all the KisTool.*::setupPainter() methods. They are not
used from now on.

I also tried to remove KisTool.*::setupPaintAction(), because the
actions should be merged in a single class (based on
KisRecorderPathPaintAction) which would handle the following operations:

paintAt
paintLine
paintPolyline
paintCurve
paintPolygon (-)
paintRect (-)
paintEllipse (-)
paintPainterPath (-)

The methods marked with (-) are currently absent in
KisRecorderPathPaintAction, so when they are implemented we can remove
setupPaintAction() functions and use KisRecorderAdepter as we already do
in KisToolFreehand.

BUG:293036