Added more options to control the PDF-output via UI.
We are re-using the KoPageLayoutWidget (which is already
translated) to allow defining the page-size (A4, Letter,
Legal, Custom, etc), the Orientation (Portrait or
Landscape) and the margins. The other two introduced
i18n-strings are already in use in calligra/libs and
therefore already translated too.

The user is able now to full control produced PDF-files
in the Desktop-UI and the calligraconverter
commandline-application.

BUG:294214