Fixed deselected selections
Now a selection cannot be "deselected". It either exists or does not exist.
When we deselect the global selection the corresponding mask is set
not-active and is left in the layer stack. When we need to "Reselect", we
just search for the last non-active mask and activate it.

These changes need testing!

BUG:294216