Fixed Deselect/Reselect to work in a "classic" way
Well, it might be probably a nice feature to be able to select the
previous selection, but the way it was done with "Reselect" might be
a bit complicated for everyday use. So just use a "classic" approach,
and take into account that it might be a cool feature to be able
continuously cycle through all the selection masks of the layer with
some hotkey.