Hide svg shape from docker
This fixes the bug reported by Mojtaba Shahi Senobari on calligra-devel, that calligra
crashes when trying to move the Embedd... shape from the add shape docker to the canvas.
It is not a crash but an ASSERT that hits and it would not crash in release mode. However
as the shape cannot be dragged to the canvas it should not be added to the docker. Therefore
we hide it now.

Please review so I can backport to 2.4

CCMAIL: calligra-devel@kde.org