Force a minimum brush spacing of 0.1
CCMAIL:kimageshop@kde.org

I notice that we still have a lot of brushes that have a default spacing
of 0, which simply doesn't perform in Krita. This commit makes the
minumum value 0.1. Big brushes as used in the Texture_paint_noisy_big
preset now actually work in real-time.

Should this be backported to 2.4?