Don't silently ignore if creating the temporary file for the content writer failed.
Related to bug 287262 but I somehow doubt that's the reason for
our famous "cannot read from beginning of file" error :-/ In
any case we should not ignore the result of the open() call
which can fail for example if the disc is full or something
else is wrong. Now we at least get a proper warning with a
hint what went wrong.

Lot of thanks goes to Stuart Dickson for the very detailed
analysis of the problem :)