Don't remove to much on removeDuplicates.
When the font-familiy is set after the removing of duplicates also keep the
properties font-pitch, font-family-generic and font-charset as these are only
evaluated if there is a font-family on the same style level.

BUG: 296667

Please review so I can backport to 2.4 branch.

Roundtrip tested with my odp and odt test set.

CCMAIL: calligra-devel@kde.org