Don't show pageX as nam of the page
in case pageX is used for saving remove that on loading. This name is used
on saving if e.g. no name for the page was set. This fixes the problem that
after save/load of a document all slides/pages are named pageX.

BUG: 294223

CCMAIL: calligra-devel@kde.org

Please review so I can backport to 2.4