Load draw:corner-radius correctly
Thr draw:corner-radius can be bigger then the actual shape is wide or high. In that case
make the radius half the width/height.

This fixes the problem that the shape in bug 260226 is shown as big as it should.

The actual problem described in the bug report had already be fixed.

Ok to backport?

BUG: 260226