Add collection to datacenters so images are saved
If there was no image in the document the image bullets where not saved as the image collection
was not added to the data centers and therefore the images where not saved. This fixes that by
adding the image collection to the datacenters.
Reorder the way the attributes are added. They need to be added before a child element is added
therefore they are moved up.
Parts of the patch are provided by gopalK.

Is it ok to include KoImageData_p.h in a lib of our own code or should we find a different solution?

Ok to backport?

CCMAIL: calligra-devel@kde.org
BUG: 299112