Fix saving of patterns
The width and the height need to be saved as length values so save as points.
When filling the style stack don't create a style that is not needed. It even
created an invalid document in case the pattern was used as a page background.
The style is added to the context once all data is filled so no need to add it
here.

Ok to backport?