Load and save the contour-polygon and contour-path elements. They are stored as KoClipPath in the engine. Previously the engine only had support for loading and saving KoClipPath in svg
Load and save the style frame style attributes that turn contoured text run around
(wrapping) on and off. wrap-contour and wrap-contour-mode

Use this new technology to do tight run around (simple change only requesting what we already had)

This commit also adds a method to pictureshape to create a tightfitting outline path of the image.
It's tested and works, but is not used yet to create a KoClipPath out of it.

REVIEW: 104873