Fix crash due to notification during file save
The reason is a notification issued in KoDocument::saveFile() which results in an async dbus message.
The reply to this message can be received (timing dependent) during deletion of the dbus conection in which case it will crash.

Also remove sheets.notifyrc file from sheets.

Review: https://git.reviewboard.kde.org/r/105010
BUG: 297822