Delayed Search: Give the user a chance to enter the whole word before begining search
When we try to search for a text that is most common in a large document the ui used to hang when the first letter was entered in the search box i.e. hung until the search for the first letter
in the whole document is over. With this commit we delay the begining of search for 1sec giving the user a chance to type the whole word he wants to search.

REVIEW: 104996