Use the rendering intent setting for displaying the image
Instead of relying on the implicit conversion inside bitBlt, which uses
INTENT_PERCEPTUAL by default, convert the bytes ourselves. This might
even give a small speedup since we call the conversion method only once
now.

It also makes it possible to use different ways of convering the
projection bytes, for instance with opencolorio.

BUG:301235