Implement most of the CPU-based OCIO path
The display filter (only one for now, and I'm not happy about the
api for setting it, it seems to me that the lcms color correction
should also be a "display filter") fitlers the pixels using ocio, which
are filtered using lcms to compensate for the monitor.

The only thing not yet done is correctly setting the ocio display
filter settings.

Even so, it doesn't seem to crash...