Show error message to the user on "Export to PDF" failure
Currently while exporting PDF we do not check if we are able to create the file at the location or not. This leads to the
"Export to PDF" action to fail silently. This patch adds check for that and the user is notified of failure.

Though I could not test in the exact environment as mentioned in bug 291925, what I did was that I created a folder with no
write permissions and tried to export PDF to that folder. In my test environment I could see that export failed silently
without this patch.

CCBUG: 291925
REVIEW: 105623