fix regression when loading vector shapes generated by OO/LO/MSO
In the commit 9efb4e33f2729080b1f1f24bb18970499a21f2b9 unfortunately a regression was
introduced so that loading of vector shapes generated by OO/LO/MSO and our filters no
longer work. Unfortunately this commit is also in the 2.5 release. This can be seen
e.g. with the following documents:

sundaram.fedorapeople.org%2Fpresentations%2Ffedora-fail-learn.odp
jeffcoweb.jeffco.k12.co.us%2Fhigh%2Fchatfield%2Fdepartments%2Fbusiness%2Fbanking_finance%2Funit_Plan_Budget.odp
drwho.virtadpt.net%2Ffiles%2FNOVALUG-Tor.odp

but much much more of our test documents are effected. There is also a crash in the
last document due to which I noticed the problem and started investigating.

The problem is that our filters and the other apps create mimetypes that se no longer
load with the above mentioned commit.

This patch fixes the filter to use the new mimetypes when converting files and adding
other mimetypes that are written be e.g. OO or guessed by the mimetype guesser of some
of the files.

Additionally fixed by this patch is to not load the first child element twice as when
the first loading failed it makes no sense to try it a second time.
It also fixes a problem that the formula shape and the chart shape tried to load every
embedded document even if it had the wrong mimetype.

REVIEW: 105873
(cherry picked from commit 4687aaf3dc93038f368f39b091ab75a40f899a07)