Fix mixup of canvas and document resource
When investigating the reason for crash for bug 304674 I found out that a very a bug had sneeked in during some refactor of the KoDocumentResourceProvider and KoCanvasResourceProvider. The problem there was that the canvas resource current page had changed and that triggered a repaint which resulted in a crash (Crash fixed by a better fix). However the reason there was a repaint was that the signal for changing the current page resource triggered in KoPathTool::resourceChanged the trigger for HandleRadius as both resource providers did trigger the same slot with oververlapping resources :-(. Seems like this was not the only place where this was done. I have fixed that by adding a second slot for the documentResourceChanged and now the different slots are used for the different resources to no longer mix stuff there. I tried my best to find all occurrences of the problem.

Will wait one week before I backport to 2.5 as asked by boemann.

REVIEW: 105896