Change QRegion to QRect in KisIndirectPaintSupport
Instead of adding a rect to a region for every dab, we coalesce the
footprints into a big rect. This makes painting faster, but at the end
of the stroke we do much more work since we bitBlt the whole united
rect to the target layer, instead of just the rects the stroke touched.

Note: we might be able to remove the mutex, too.