Fixed jumping while mirror+zoom in vastScrolling mode
It's quite a pity that this patch adds quite weird small bugs to the
non-vastScrolling mode. These bugs can be caught by the disabled tests
in KisZoomAndPanTest. The problem is that to fix these bugs I'd have to
make really drastic changes to the KoCanvasControllerWidget (including
the change of the external interface). The more detailed description is
give in the test itself:

"We have to disable a couple of tests here for the case when
vastScrolling value is 0.2. The problem is that the centering
correction applied  to the offset in
KisCanvasController::rotateCanvas pollutes the preferredCenter
value, because KoCnvasControllerWidget has no access to this
correction and cannot calculate the real value of the center of
the image. To fix this bug the calculation of correction
(aka "origin") should be moved to the KoCanvasControllerWidget
itself which would cause quite huge changes (including the change
of the external interface of it). Namely, we would have to
*calculate* offset from the value of the scroll bars, but not
use their values directly:

offset = scrollBarValue - origin

So now we just disable these unittests and allow a couple
of "jumping" bugs appear in vastScrolling < 0.5 modes, which
is, actually, not the default case."