Fix filter selection dialog
Of course, when this dialog pops up, it's too late anyway, but that's
no reason to let it be buggy: it used a stringlist and a listwidget.
The list widget's contents got sorted, changing the order. Then the current
item's index was used to index the stringlist, which meant we would return
the wrong mimetype. Store the mimetype as user data in the listwidgetitem
instead.