Use KoBorder to load the borders.
This gains us support of loading double borders. And since ms type borders
was ported to KoBorder in previous commit we retain loading and gain saving for that as well.

Drawback is that we now loose loading parentstyle borders. In the long run this is
correct as tablecellstyle should have inheritance like para and char styles, but for
now we have lost it. Good news is that I've never seen any odf file that uses this in
practice.

BUG:305932