Added a PACKAGERS_BUILD option for code generation for many architectures
This option is disabled by default.

By default we build the whole Calligra optimized for the host architecture.
When the option is on, the hottest parts of calligra will compile optimized
for several most popular architectures. The rest of the code will not use
any brand-new instructions for not breaking binary compatibility among cpus.

Short manual:
1) If you build Calligra for yourself and are not going to copy Krita binary
   to another CPU, disable this option.
2) If you build a Calligra package and are going to distribute it among users,
   then enable the option.