Make the textRangesChangingWithin method actually check all that it's supposed to check
Normally we don't use this method in a too complex way so there should be no regressions,
but still if bugs come up with saving too few textranges, this will be a patch to investigate

Basically we were only checking the matching tag against one of the parameters while we should be
checking against both. I most case the our use is to have global match limits, so that is why we
have not noticed the bug.

REVIEW: 107621

thanks Friedrich for reviewing