Added two classes for the new version of KisTransformWorker
KisFixedPoint now represents the .8 fixed point arithmetics. Current
hardcoded fp.8 implementation in the KisTransformWorker has
various flaws like rounding and dealing with negative values.

KisFilterWeightsBuffer represents a part of KisTransformWorker that
generates weights for the pixel blending. Being able to add it to
the unittest I could catch many mathematical problems in it.

Now these is one more class to be written (that will do the actual
blending) and then it all will have to be merged into KisTransformWorker.

CCBUG:289645